fix mistakes in static library

+++ lib_static/include/lib_static.h
@@ -1,8 +1,8 @@
 #ifndef IZ_2_LIB_STATIC_H
 #define IZ_2_LIB_STATIC_H
 
-#include <stdint.h>
+#include <stddef.h>
 
-uint32_t countRPeaks(const int* records, uint32_t size, uint16_t R_window);
+size_t countRPeaks(const int* records, size_t size, size_t R_window);
 
 #endif // IZ_2_LIB_STATIC_H
